# Header inclusion cleanup