--- a/es-app/src/FileData.h
+++ b/es-app/src/FileData.h
@@ -2,13 +2,12 @@
 #ifndef ES_APP_FILE_DATA_H
 #define ES_APP_FILE_DATA_H
 
-#include <unordered_map>
-#include <string>
-#include <vector>
-#include <boost/filesystem.hpp>
 #include "MetaData.h"
+#include <boost/filesystem/path.hpp>
+#include <unordered_map>
 
 class SystemData;
+class Window;
 struct SystemEnvironmentData;
 
 enum FileType