--- a/es-app/src/FileSorts.h
+++ b/es-app/src/FileSorts.h
@@ -2,9 +2,8 @@
 #ifndef ES_APP_FILE_SORTS_H
 #define ES_APP_FILE_SORTS_H
 
-#include <vector>
 #include "FileData.h"
-#include "SystemData.h"
+#include <vector>
 
 namespace FileSorts
 {