--- a/es-app/src/PlatformId.cpp
+++ b/es-app/src/PlatformId.cpp
@@ -1,4 +1,5 @@
 #include "PlatformId.h"
+
 #include <string.h>
 
 extern const char* mameNameToRealName[];