--- a/es-app/src/VolumeControl.h
+++ b/es-app/src/VolumeControl.h
@@ -3,7 +3,6 @@
 #define ES_APP_VOLUME_CONTROL_H
 
 #include <memory>
-#include <stdint.h>
 
 #if defined (__APPLE__)
     #error TODO: Not implemented for MacOS yet!!!
@@ -13,9 +12,8 @@
 	#include <alsa/asoundlib.h>
 #elif defined(WIN32) || defined(_WIN32)
 	#include <Windows.h>
-	#include <MMSystem.h>
-	#include <mmdeviceapi.h>
 	#include <endpointvolume.h>
+	#include <mmeapi.h>
 #endif
 
 /*!