--- a/es-app/src/animations/LaunchAnimation.h
+++ b/es-app/src/animations/LaunchAnimation.h
@@ -3,7 +3,6 @@
 #define ES_APP_ANIMATIONS_LAUNCH_ANIMATION_H
 
 #include "animations/Animation.h"
-#include "Log.h"
 
 // let's look at the game launch effect:
 // -move camera to center on point P (interpolation method: linear)