--- a/es-app/src/guis/GuiCollectionSystemsOptions.h
+++ b/es-app/src/guis/GuiCollectionSystemsOptions.h
@@ -2,17 +2,12 @@
 #ifndef ES_APP_GUIS_GUI_COLLECTION_SYSTEM_OPTIONS_H
 #define ES_APP_GUIS_GUI_COLLECTION_SYSTEM_OPTIONS_H
 
-#include "GuiComponent.h"
-#include "SystemData.h"
 #include "components/MenuComponent.h"
-#include "CollectionSystemManager.h"
-#include "components/SwitchComponent.h"
-#include "Log.h"
-
 
 template<typename T>
 class OptionListComponent;
-
+class SwitchComponent;
+class SystemData;
 
 class GuiCollectionSystemsOptions : public GuiComponent
 {