--- a/es-app/src/guis/GuiGameScraper.h
+++ b/es-app/src/guis/GuiGameScraper.h
@@ -2,9 +2,9 @@
 #ifndef ES_APP_GUIS_GUI_GAME_SCRAPER_H
 #define ES_APP_GUIS_GUI_GAME_SCRAPER_H
 
-#include "GuiComponent.h"
-#include "components/ScraperSearchComponent.h"
 #include "components/NinePatchComponent.h"
+#include "components/ScraperSearchComponent.h"
+#include "GuiComponent.h"
 
 class GuiGameScraper : public GuiComponent
 {