--- a/es-app/src/guis/GuiGeneralScreensaverOptions.h
+++ b/es-app/src/guis/GuiGeneralScreensaverOptions.h
@@ -2,7 +2,6 @@
 #ifndef ES_APP_GUIS_GUI_GENERAL_SCREENSAVER_OPTIONS_H
 #define ES_APP_GUIS_GUI_GENERAL_SCREENSAVER_OPTIONS_H
 
-#include "components/MenuComponent.h"
 #include "GuiScreensaverOptions.h"
 
 class GuiGeneralScreensaverOptions : public GuiScreensaverOptions