--- a/es-app/src/guis/GuiInfoPopup.cpp
+++ b/es-app/src/guis/GuiInfoPopup.cpp
@@ -1,7 +1,10 @@
 #include "guis/GuiInfoPopup.h"
-#include "Renderer.h"
+
+#include "components/ComponentGrid.h"
+#include "components/NinePatchComponent.h"
 #include "components/TextComponent.h"
-#include "Log.h"
+#include "Renderer.h"
+#include <SDL_timer.h>
 
 GuiInfoPopup::GuiInfoPopup(Window* window, std::string message, int duration) :
 	GuiComponent(window), mMessage(message), mDuration(duration), running(true)