--- a/es-app/src/guis/GuiInfoPopup.h
+++ b/es-app/src/guis/GuiInfoPopup.h
@@ -3,12 +3,10 @@
 #define ES_APP_GUIS_GUI_INFO_POPUP_H
 
 #include "GuiComponent.h"
-#include "components/NinePatchComponent.h"
-#include "components/ComponentGrid.h"
 #include "Window.h"
-#include "Log.h"
 
-
+class ComponentGrid;
+class NinePatchComponent;
 
 class GuiInfoPopup : public GuiComponent, public Window::InfoPopup
 {