--- a/es-app/src/guis/GuiMenu.h
+++ b/es-app/src/guis/GuiMenu.h
@@ -2,9 +2,8 @@
 #ifndef ES_APP_GUIS_GUI_MENU_H
 #define ES_APP_GUIS_GUI_MENU_H
 
-#include "GuiComponent.h"
 #include "components/MenuComponent.h"
-#include <functional>
+#include "GuiComponent.h"
 
 class GuiMenu : public GuiComponent
 {