--- a/es-app/src/guis/GuiScraperMulti.cpp
+++ b/es-app/src/guis/GuiScraperMulti.cpp
@@ -1,15 +1,15 @@
 #include "guis/GuiScraperMulti.h"
-#include "Renderer.h"
-#include "Log.h"
+
+#include "components/ButtonComponent.h"
+#include "components/MenuComponent.h"
+#include "components/ScraperSearchComponent.h"
+#include "components/TextComponent.h"
+#include "guis/GuiMsgBox.h"
 #include "views/ViewController.h"
 #include "Gamelist.h"
 #include "PowerSaver.h"
-
-#include "components/TextComponent.h"
-#include "components/ButtonComponent.h"
-#include "components/ScraperSearchComponent.h"
-#include "components/MenuComponent.h" // for makeButtonGrid
-#include "guis/GuiMsgBox.h"
+#include "SystemData.h"
+#include "Window.h"
 
 GuiScraperMulti::GuiScraperMulti(Window* window, const std::queue<ScraperSearchParams>& searches, bool approveResults) :
 	GuiComponent(window), mBackground(window, ":/frame.png"), mGrid(window, Vector2i(1, 5)),