--- a/es-app/src/guis/GuiScraperStart.cpp
+++ b/es-app/src/guis/GuiScraperStart.cpp
@@ -1,11 +1,12 @@
 #include "guis/GuiScraperStart.h"
-#include "guis/GuiScraperMulti.h"
-#include "guis/GuiMsgBox.h"
-#include "views/ViewController.h"
 
-#include "components/TextComponent.h"
 #include "components/OptionListComponent.h"
 #include "components/SwitchComponent.h"
+#include "guis/GuiMsgBox.h"
+#include "guis/GuiScraperMulti.h"
+#include "views/ViewController.h"
+#include "FileData.h"
+#include "SystemData.h"
 
 GuiScraperStart::GuiScraperStart(Window* window) : GuiComponent(window),
 	mMenu(window, "SCRAPE NOW")