--- a/es-app/src/guis/GuiSettings.cpp
+++ b/es-app/src/guis/GuiSettings.cpp
@@ -1,7 +1,9 @@
 #include "guis/GuiSettings.h"
-#include "Window.h"
-#include "Settings.h"
+
 #include "views/ViewController.h"
+#include "Settings.h"
+#include "SystemData.h"
+#include "Window.h"
 
 GuiSettings::GuiSettings(Window* window, const char* title) : GuiComponent(window), mMenu(window, title)
 {