--- a/es-app/src/scrapers/Scraper.cpp
+++ b/es-app/src/scrapers/Scraper.cpp
@@ -1,11 +1,15 @@
 #include "scrapers/Scraper.h"
-#include "Log.h"
-#include "Settings.h"
-#include <FreeImage.h>
-#include <boost/filesystem.hpp>
-#include <boost/assign.hpp>
 
+#include "FileData.h"
 #include "GamesDBScraper.h"
+#include "Log.h"
+#include "platform.h"
+#include "Settings.h"
+#include "SystemData.h"
+#include <boost/assign/list_of.hpp>
+#include <boost/filesystem/operations.hpp>
+#include <FreeImage.h>
+#include <fstream>
 
 const std::map<std::string, generate_scraper_requests_func> scraper_request_funcs = boost::assign::map_list_of
 	("TheGamesDB", &thegamesdb_generate_scraper_requests);