… es-app/src/guis/GuiMenu.cpp
… es-app/src/views/SystemView.cpp
… es-app/src/views/SystemView.h
… es-app/src/views/ViewController.cpp
… es-app/src/views/ViewController.h
… es-app/src/views/gamelist/BasicGameListView.cpp
… es-app/src/views/gamelist/BasicGameListView.h
… es-app/src/views/gamelist/DetailedGameListView.cpp
… es-app/src/views/gamelist/DetailedGameListView.h
… es-app/src/views/gamelist/GridGameListView.cpp
… es-app/src/views/gamelist/GridGameListView.h
… es-app/src/views/gamelist/IGameListView.cpp
… es-app/src/views/gamelist/IGameListView.h
… es-app/src/views/gamelist/ISimpleGameListView.cpp
… es-app/src/views/gamelist/ISimpleGameListView.h
… es-app/src/views/gamelist/VideoGameListView.cpp
… es-app/src/views/gamelist/VideoGameListView.h
… es-core/src/AudioManager.cpp
… es-core/src/AudioManager.h
… es-core/src/GuiComponent.cpp
… es-core/src/GuiComponent.h
… es-core/src/HelpPrompt.h
… es-core/src/HelpStyle.cpp
… es-core/src/HelpStyle.h
… es-core/src/HttpReq.cpp
… es-core/src/HttpReq.h
… es-core/src/ImageIO.cpp
… es-core/src/ImageIO.h
… es-core/src/InputConfig.cpp
… es-core/src/InputConfig.h