Basically removing / moving includes, and replacing with forward declarations where possible.
This could potentially be improved even more if I went over all files one more time, but I doubt it would be a noticeable difference.

Only code change is in es-app/src/FileFilterIndex.cpp where I replaced:
`ratingNumber = boost::math::iround(std::stod(ratingString)*5);`
with:
`ratingNumber = (int)((std::stod(ratingString)*5)+0.5);`
to avoid a boost header for just that simple round.

I also moved the HelpPrompt typedef to a new file: es-core/src/HelpPrompt.h
to avoid having to include GuiComponent in other headers that only needed to know about the HelpPrompt.

Now for some statistics!
Header Hero is a  tool that estimates how many lines are parsed by all header inclusions, but it doesn't do real preprocessing so it's just a rough estimate.
The times are for a fresh recompile of the entire thing.

```
Header Hero:
    master:  36 905 936
    tomaz82: 20 838 000        ( ~44% less )

Windows 7 MSVS 2015:
    master:   1m 09s (   69s )
    tomaz82:  0m 46s (   46s ) ( ~33% less )

VMPlayer Ubuntu 16.04:
    master:   3m 35s (  215s )
    tomaz82:  2m 13s (  133s ) ( ~38% less )

rpi3:
    master:  34m 14s ( 2054s )
    tomaz82: 19m 28s ( 1168s ) ( ~43% less )
```

Now this is just the second step of my master plan for world domina... I mean, for cleaning up and optimizing ES, the next step will be more intrusive since I'm gonna need to change a lot of method parameters. But that's a later worry.