@zigurana 2017-11-02
So, this is touching a lot of files. Did you use some kind of tool to clean up these headers, or did you go over them one-by-one?