@zigurana 2017-11-02
Not to distract you from your path to world domination (via retro-games no less!), but would you mind sharing what your plans for further optimizing ES are?
More specifically, I am working on two (largish) refactor activities (#247 and #252), and I am wondering how well that would mesh with your plans? This more so because you seem to be working at a **much** higher productivity level than I am able to.