@tomaz82 2017-11-02
@zigurana No, not a lot of files, all files :P And no tools, all by hand. Remove all header includes, compile, add back the ones needed to get it to compile again. Move onto the next file.

My goal with the next step is to make all method parameters that can, be references, right now ES is copying a lot of data around since it doesn't use references and pointers properly. And doing that will allow me to clean up the header inclusions even further.